Update module-2-homework.markdown
The link to submit this weekend's hw (in Day 4) still points to 1510. I created a new list for 1602 and created a pull request for it. This pull request links to the pull request for the new 1602 sheet, so if you accept that new sheet then you'll need to update the link in this doc as well.

I hope that doesn't create more work than it saves.